ieee1588: Remove some implicit conversions
From teensy41.

+++ src/qnethernet/drivers/driver_teensy41.c
@@ -693,7 +693,7 @@ static struct pbuf* low_level_input(volatile enetbufferdesc_t* const pBD) {
           // The timer has wrapped around
           --p->timestamp.tv_sec;
         }
-        p->timestamp.tv_nsec = pBD->timestamp;
+        p->timestamp.tv_nsec = (long)pBD->timestamp;
       }
     } else {
       LINK_STATS_INC(link.drop);
@@ -737,7 +737,7 @@ static inline void update_bufdesc(volatile enetbufferdesc_t* const pBD,
     s_doTimestampNext = false;
     pBD->extend1 |= kEnetTxBdTimestamp;
   } else {
-    pBD->extend1 &= ~kEnetTxBdTimestamp;
+    pBD->extend1 &= (uint16_t)(~kEnetTxBdTimestamp);
   }
 
   ENET_TDAR = ENET_TDAR_TDAR;
@@ -785,7 +785,7 @@ static void enet_isr(void) {
   if ((ENET_EIR & ENET_EIR_TS_AVAIL) != 0) {
     ENET_EIR = ENET_EIR_TS_AVAIL;
     s_txTimestamp.tv_sec = s_ieee1588Seconds;
-    s_txTimestamp.tv_nsec = ENET_ATSTMP;
+    s_txTimestamp.tv_nsec = (long)ENET_ATSTMP;
     s_hasTxTimestamp = true;
   }
 
@@ -1421,7 +1421,7 @@ bool driver_ieee1588_read_timer(struct timespec* const t) {
   while ((ENET_ATCR & ENET_ATCR_CAPTURE) != 0) {
     // Wait for bit to clear
   }
-  t->tv_nsec = ENET_ATVR;
+  t->tv_nsec = (long)ENET_ATVR;
 
   // The timer could have wrapped while we were doing stuff
   // Leave the interrupt set so that our internal timer will catch it
@@ -1439,8 +1439,8 @@ bool driver_ieee1588_write_timer(const struct timespec* const t) {
   }
 
   qnethernet_hal_disable_interrupts();  // {
-  s_ieee1588Seconds = t->tv_sec;
-  ENET_ATVR = t->tv_nsec;
+  s_ieee1588Seconds = (uint32_t)t->tv_sec;
+  ENET_ATVR = (uint32_t)t->tv_nsec;
   qnethernet_hal_enable_interrupts();  // }
 
   return true;
@@ -1493,7 +1493,7 @@ bool driver_ieee1588_adjust_freq(int nsps) {
     // Speed up
     ++inc;
   }
-  return driver_ieee1588_adjust_timer(inc, F_ENET_TS_CLK / nsps);
+  return driver_ieee1588_adjust_timer(inc, (uint32_t)(F_ENET_TS_CLK / nsps));
 }
 
 // Channels
